requirements.txt
bcrypt==3.1.0
beautifulsoup4==4.5.1
cffi==1.7.0
ecdsa==0.13
falcon==1.0.0
falcon-cors==1.0.1
future==0.15.2
gunicorn==19.6.0
peewee==2.8.3
pycparser==2.14
pycrypto==2.6.1
python-jose==1.3.2
python-mimeparse==1.5.2
six==1.10.0
waitress==1.0.0
WebOb==1.6.2
WebTest==2.0.23
